Generate Syncthing ConfigClass using xsd2code.
