+++ SyncthingTray/External/GenerateSyncthingConfigClass.cmd
@@ -1,0 +1,2 @@
+rem xsd.exe SyncThingConfig.xsd /classes /n:SyncthingTray
+xsd2code.exe SyncThingConfig.xsd SyncthingTray.External /is /ap+ /cl+
